lots of bugs fixed, new css, new html, new id's and classes, and ideas to add feedback button and 'who we are' at index.html to no logged in users

+++ account.js
@@ -1,8 +1,9 @@
+// Load website user data
 firebase.auth().onAuthStateChanged(function(user) {
 	if (user) {
-		// User is signed in, so show the loggedin-div
-		document.getElementById("login-div").style.display = "none";
-		document.getElementById("loggedin-div").style.display = "block";
+		// User is logged in, so show the account
+		document.getElementById('log').style.display = 'none';
+		document.getElementById('account').style.display = 'flex';
 
 		var name = user.displayName;
 		var email = user.email;
@@ -12,86 +13,70 @@ firebase.auth().onAuthStateChanged(function(user) {
 
 		// Welcome message
 		if (name != null) {
-			document.getElementById("user-welcome").innerHTML = "Bienvenido/a " + name;
+			document.getElementById('user-welcome').innerHTML = 'Bienvenido/a ' + name;
 		} else {
-			document.getElementById("user-welcome").innerHTML = "Bienvenido/a " + email;
+			document.getElementById('user-welcome').innerHTML = 'Bienvenido/a ' + email;
 		}
 		
-		// Imprimir datos de perfil (Auth)
+		// Show profile data (Auth)
 		if (user.photoURL != null){
-			document.getElementById("profile-image").innerHTML = "<img src=\"" + photoURL + "\">";
+			document.getElementById('account-image').innerHTML = '<img src="' + user.photoURL + '">';
 		} else {
 			// In case there was an image before and the user has 
 			// removed it. It must go away.
-			document.getElementById("profile-image").innerHTML = null;
+			document.getElementById('account-image').innerHTML = null;
 		}
-		document.getElementById("name-input").value = name;
-		document.getElementById("email-data").innerHTML = email;
-		document.getElementById("photourl-input").value = user.photoURL;
+		document.getElementById('account-name-input').value = name;
+		document.getElementById('account-email-data').innerHTML = email;
+		document.getElementById('account-imagelink-input').value = user.photoURL;
 
 		// Show DB data
-		/*
-		//Realtime
-		firebase.database().ref('users/' + user.uid + '/bio').on('value',function(snapshot){
-			document.getElementById("bio-input").value = snapshot.val();
-		});
-		*/
 		firebase.database().ref('users/' + user.uid).once('value').then(function(snapshot){
-			// If the user exists at the DB... 
+			// If the user exists at the DB... (snapshot.val())
 			// (otherwise wait the register function to create it. It must be first
 			// authenticated)
 			if(snapshot.val() != null){
 				if (snapshot.val().username != undefined){
-					document.getElementById("username-input").value = snapshot.val().username;
+					document.getElementById('account-username-input').value = snapshot.val().username;
 				}
 
 				if (snapshot.val().surname != undefined){
-					document.getElementById("surname-input").value = snapshot.val().surname;
+					document.getElementById('account-surname-input').value = snapshot.val().surname;
 				}
 
 				if (snapshot.val().bio != undefined){
-					document.getElementById("bio-input").value = snapshot.val().bio;
+					document.getElementById('account-bio-input').value = snapshot.val().bio;
 				}
 
 				if (snapshot.val().freetext != undefined){
-					document.getElementById("freetext-input").value = snapshot.val().freetext;
+					document.getElementById('account-freetext-input').value = snapshot.val().freetext;
 				}
 				
-				if (user.emailVerified == false){
-					document.getElementById("verify-email").innerHTML = '<button onclick="verifyEmail()">Verificar email</button>';
+				if (user.emailVerified == true){
+					document.getElementById('verify-email').style.display = 'none';
+				} else {
+					document.getElementById('verify-email').style.display = 'flex';
 				}
 				
 				if (snapshot.val().birthdate != undefined){
-					document.getElementById("birthdate-input").value = snapshot.val().birthdate;
+					document.getElementById('account-birthday-input').value = snapshot.val().birthdate;
 				}
 				// TODO: MAKE THIS SHORT AND ELEGANT
 			};
 		});
 
 	} else {
-		// No user is signed in.
-		document.getElementById("login-div").style.display = "block";
-		document.getElementById("loggedin-div").style.display = "none";
+		// No user is logged in. (If recently logged out, and for redundancy)
+		document.getElementById('log').style.display = 'flex';
+		document.getElementById('account').style.display = 'none';
 	}
 });
 
 
-function login(){
-	var email = document.getElementById("email-input-login").value;
-	var pwd = document.getElementById("pwd-input-login").value;
-
-	firebase.auth().signInWithEmailAndPassword(email, pwd).catch(function(error) {
-		// Handle Errors here.
-		var errorCode = error.code;
-		var errorMessage = error.message;
-
-		alert("Error: " + errorMessage);
-	});
-}
 function register(){
 
-	var email = document.getElementById("email-input-login").value;
-	var pwd = document.getElementById("pwd-input-login").value;
+	var email = document.getElementById('email-input-login').value;
+	var pwd = document.getElementById('pwd-input-login').value;
 
 	firebase.auth().createUserWithEmailAndPassword(email, pwd).catch(function(error) {
 		alert('Error: ' + error.message);
@@ -100,7 +85,20 @@ function register(){
 	// Just after registry, the onAuthStateChanged function will try to show data that 
 	// still doesn't exist in the DB, and will show an error. Wait 100ms to have a 
 	// registered user, and then update the data to the DB.
-	setTimeout(update, 5000);
+	setTimeout(update, 1000);
+}
+function login(){
+	var email = document.getElementById('email-input').value;
+	var pwd = document.getElementById('pwd-input').value;
+
+	firebase.auth().signInWithEmailAndPassword(email, pwd).catch(function(error) {
+		// Handle Errors here.
+		alert('Error: ' + errorMessage);
+	});
+
+	// (Don't keep the input data in the html)
+	email = null;
+	pwd = null;
 }
 function logout(){
 	firebase.auth().signOut();
@@ -111,18 +109,10 @@ function google_login(){
 	firebase.auth().signInWithPopup(provider).then(function(result) {
 		// This gives you a Google Access Token. You can use it to access the Google API.
 		var token = result.credential.accessToken;
-		// The signed-in user info.
+		// The logged-in user info.
 		var user = result.user;
-		// ...
+
 	}).catch(function(error) {
-		// Handle Errors here.
-		var errorCode = error.code;
-		var errorMessage = error.message;
-		// The email of the user's account used.
-		var email = error.email;
-		// The firebase.auth.AuthCredential type that was used.
-		var credential = error.credential;
-		// ...
 		alert('Error: ' + error.message);
 		console.log(error);
 	});
@@ -133,19 +123,12 @@ function google_login(){
 // 	firebase.auth().signInWithPopup(provider).then(function(result) {
 // 		// This gives you a Google Access Token. You can use it to access the Google API.
 // 		var token = result.credential.accessToken;
-// 		// The signed-in user info.
+// 		// The logged-in user info.
 // 		var user = result.user;
 // 		// ...
 // 	}).catch(function(error) {
-// 		// Handle Errors here.
-// 		var errorCode = error.code;
-// 		var errorMessage = error.message;
-// 		// The email of the user's account used.
-// 		var email = error.email;
-// 		// The firebase.auth.AuthCredential type that was used.
-// 		var credential = error.credential;
-// 		// ...
-// 		alert(error);
+// 		alert('Error: ' + error.message);
+// 		console.log(error);
 // 	});
 // }
 function passwordForgotten(){
@@ -153,18 +136,18 @@ function passwordForgotten(){
 	var auth = firebase.auth();
 
 	auth.sendPasswordResetEmail(email).then(function() {
-		alert("Email sent");
+		alert('Email sent');
 	}).catch(function(error) {
 		alert('Error: ' + error.message);
 		console.log(error);
 	});
 }
 function resetPassword(){
-	var email = firebase.auth().currentUser.email;
 	var auth = firebase.auth();
+	var email = auth.currentUser.email;
 
 	auth.sendPasswordResetEmail(email).then(function() {
-		alert("Email sent");
+		alert('Email sent');
 	}).catch(function(error) {
 		alert('Error: ' + error.message);
 		console.log(error);
@@ -172,7 +155,7 @@ function resetPassword(){
 }
 function verifyEmail(){
 	firebase.auth().currentUser.sendEmailVerification().then(function(){
-		alert("Email sent");
+		alert('Email sent');
 	}).catch(function(error){
 		alert('Error: ' + error.message);
 		console.log(error);
@@ -180,43 +163,46 @@ function verifyEmail(){
 }
 
 function deleteAccount() {
-	var user = firebase.auth().currentUser;
+	var rlly = confirm('¿Estás seguro de que quieres borrar la cuenta?');
+	if(rlly){
+		var user = firebase.auth().currentUser;
 
-	user.delete().then(function() {
-		// User deleted.
-	}).catch(function(error) {
-		alert('Error: ' + error.message);
-		console.log(error);
-	});
+		user.delete().then(function() {
+			// User deleted.
+		}).catch(function(error) {
+			alert('Error: ' + error.message);
+			console.log(error);
+		});
+	}
 }
 
 
 function update(){
+	console.log('Updating data...');
 	var user = firebase.auth().currentUser;
 	
 	// Update auth data (name and photoURL)
 	user.updateProfile({
-		displayName: document.getElementById("name-input").value,
-		photoURL: document.getElementById("photourl-input").value
+		displayName: document.getElementById('account-name-input').value,
+		photoURL: document.getElementById('account-imagelink-input').value
 	}).then(function() {
 		// Update successful.
-		// alert("Update successful");
+		// alert('Update successful');
 	}).catch(function(error) {
 		// An error happened.
 		// alert(error);
 	});
 
-	// console.log('updating data...' + user.uid);
 	// Profile data to DB
 	firebase.database().ref('users/' + user.uid).set({
 		email: user.email,
-		username: document.getElementById("username-input").value,
+		username: document.getElementById('account-username-input').value,
 		name: user.displayName,
-		surname: document.getElementById("surname-input").value,
+		surname: document.getElementById('account-surname-input').value,
 		photo_url: user.photoURL,
 		emailVerified: user.emailVerified, // Not secure
-		bio: document.getElementById("bio-input").value,
-		freetext: document.getElementById("freetext-input").value,
-		birthdate: document.getElementById("birthdate-input").value
+		bio: document.getElementById('account-bio-input').value,
+		freetext: document.getElementById('account-freetext-input').value,
+		birthdate: document.getElementById('account-birthday-input').value
 	});
 }
